# app/api/v1/trf/route.js
import { NextResponse } from 'next/server';
import { getTRFList } from '@/lib/services/analyticsService';

export const dynamic = 'force-dynamic';

const CACHE_HEADERS = {
    'Cache-Control': 'public, max-age=3600, s-maxage=86400, stale-while-revalidate=604800',
    'Netlify-Vary': 'query',
    'Access-Control-Allow-Origin': '*',
};

export async function GET(request) {
    try {
        const { searchParams } = new URL(request.url);
        const district = searchParams.get('district') || '';
        const zone = searchParams.get('zone') || '';
        const base = searchParams.get('base') || '';
        const country = searchParams.get('country') || '';
        const sortBy = searchParams.get('sortBy') || searchParams.get('sort_by') || 'totalContributionsUSD';
        const sortOrder = searchParams.get('sortOrder') || searchParams.get('sort_order') || 'desc';
        const limit = searchParams.get('limit') || 25;
        const offset = searchParams.get('offset') || 0;

        const result = await getTRFList({
            district,
            zone,
            base,
            country,
            sortBy,
            sortOrder,
            limit,
            offset
        });

        return NextResponse.json(result, { headers: CACHE_HEADERS });
    } catch (error) {
        console.error('API /v1/trf error:', error);
        return NextResponse.json({ error: 'Failed to retrieve TRF contributions' }, { status: 500 });
    }
}