# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Rotaract South Asia Analytics API",
    "version": "1.1.0",
    "description": "Public, high-performance REST API and Model Context Protocol (MCP) server providing open analytics data for Rotaract and Rotary across South Asia (Zones 4, 5, 6, and 7; 44 districts; 2,870+ clubs).\n\n### Model Context Protocol (MCP) Integration\nThis service includes a Model Context Protocol (MCP) server for direct use by AI assistants (Cursor, Claude Desktop, Antigravity, VS Code):\n- **Local Stdio Command:** `npm run mcp` or `node scripts/mcp_server.js`\n- **Remote HTTP Endpoint:** `POST https://insights.rsmda.org/api/mcp` (JSON-RPC 2.0)\n- **Available AI Tools (14 tools):** `get_summary`, `get_leaderboards`, `get_districts`, `get_district_insights`, `get_zone_summary`, `search_clubs`, `get_club_profile`, `get_new_clubs`, `find_compliance_risks`, `find_dual_risk_clubs`, `get_interact_analytics`, `find_rotary_opportunities`, `get_foundation_giving`, `get_worldwide_rankings`.\n\nAll endpoints are open, unauthenticated, and CORS-enabled.",
    "contact": {
      "name": "Rotaract South Asia Analytics Team",
      "url": "https://insights.rsmda.org"
    },
    "license": {
      "name": "Open Data Commons Open Database License (ODbL)",
      "url": "https://opendatacommons.org/licenses/odbl/"
    }
  },
  "servers": [
    {
      "url": "https://insights.rsmda.org",
      "description": "Production server (Primary)"
    },
    {
...
    },
    {
...
    }
...
  "paths": {
...
    "/api/v1/leaderboards": {
      "get": {
        "tags": [
          "Leaderboards"
        ],
        "summary": "Ranked Leaderboards & Top Charts",
        "description": "Returns top rankings across South Asia for largest clubs (by membership), top TRF donors, highest arrears dues, and top districts. Supports optional filtering by district, zone, base, and category.",
        "operationId": "getLeaderboards",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": false,
            "description": "Leaderboard category ('all', 'largest_clubs', 'community_clubs', 'university_clubs', 'trf_giving', 'highest_arrears', 'districts_by_clubs', 'districts_by_members', 'districts_by_trf', 'districts_by_arrears', 'districts_by_no_officers', 'new_clubs')",
            "schema": {
              "type": "string",
              "default": "all"
            }
          },
          {
            "name": "district",
            "in": "query",
            "required": false,
            "description": "Scope rankings to a specific 4-digit district (e.g. '3000')",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "zone",
            "in": "query",
            "required": false,
            "description": "Scope rankings to a specific zone (e.g. 'Zone 5' or '5')",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scope",
            "in": "query",
            "required": false,
            "description": "Geographic scope: 'south_asia' (default, Zones 4, 5, 6, 7) or 'worldwide' (for global rankings)",
            "schema": {
              "type": "string",
              "enum": [
                "south_asia",
                "worldwide"
              ],
              "default": "south_asia"
            }
          },
          {
            "name": "base",
            "in": "query",
            "required": false,
            "description": "Filter by base ('Community' or 'University')",
            "schema": {
              "type": "string",
              "enum": [
                "Community",
                "University"
              ]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Number of top records to return (1-50, default 10)",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "description": "Number of records to skip for pagination",
            "schema": {
              "type": "integer",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ranked leaderboards data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeaderboardsResponse"
                },
                "example": {
                  "category": "largest_clubs",
                  "count": 3,
                  "offset": 0,
                  "limit": 3,
                  "data": [
                    {
                      "rank": 1,
                      "id": "8825786",
                      "name": "MGR University",
                      "district": "3233",
                      "zone": "Zone 5",
                      "country": "India",
                      "base": "University",
                      "members": 2988,
                      "sponsorClubs": "Chennai Metropolis (Rotary)"
                    },
                    {
                      "rank": 2,
                      "id": "8828540",
                      "name": "Jayaraj Annapackiam College for Women (Autonomous)",
                      "district": "3000",
                      "zone": "Zone 5",
                      "country": "India",
                      "base": "University",
                      "members": 2119,
                      "sponsorClubs": "Theni United Inspirers (Rotary)"
                    },
                    {
                      "rank": 3,
                      "id": "8828452",
                      "name": "Trichy SRM Medical College Hospital and Research Centre",
                      "district": "3000",
                      "zone": "Zone 5",
                      "country": "India",
                      "base": "University",
                      "members": 1854,
                      "sponsorClubs": "Trichy One Health (Rotary)"
                    }
                  ]
                }
              }
            }
          }
        },
        "x-openai-isConsequential": false,
        "x-ms-require-user-confirmation": false
      }
    },
...
    "/api/v1/opportunities/rotary": {
      "get": {
        "tags": [
          "Opportunities"
        ],
        "summary": "Rotary Sponsorship Opportunities",
        "description": "Identifies Rotary clubs without sponsored Rotaract clubs or without Interact clubs.",
        "operationId": "getRotaryOpportunities",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "required": false,
            "description": "Type of opportunity: 'no_rotaract' (default) or 'no_interact'",
            "schema": {
              "type": "string",
              "enum": [
                "no_rotaract",
                "no_interact"
              ],
              "default": "no_rotaract"
            }
          },
          {
            "name": "district",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "zone",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 25
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "required": false,
            "description": "Sort field (\"district\", \"name\", \"zone\", \"members\")",
            "schema": {
              "type": "string",
              "enum": [
                "district",
                "name",
                "zone",
                "members"
              ],
              "default": "district"
            }
          },
          {
            "name": "sortOrder",
            "in": "query",
            "required": false,
            "description": "Sort order (\"asc\" or \"desc\")",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ],
              "default": "asc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of Rotary opportunity clubs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RotaryOpportunityResponse"
                }
              }
            }
          }
        },
        "x-openai-isConsequential": false,
        "x-ms-require-user-confirmation": false
      }
    },
    "/api/v1/trf": {
      "get": {
        "tags": [
          "Foundation"
        ],
        "summary": "The Rotary Foundation (TRF) Contributions",
        "description": "Retrieves giving statistics, Annual Fund contributions, and PolioPlus donations per club.",
        "operationId": "getTRFContributions",
        "parameters": [
          {
            "name": "district",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "zone",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 25
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "country",
            "in": "query",
            "required": false,
            "description": "Optional country filter (e.g. \"India\", \"Nepal\", \"Sri Lanka\")",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "base",
            "in": "query",
            "required": false,
            "description": "Optional club base filter (\"Community\" or \"University\")",
            "schema": {
              "type": "string",
              "enum": [
                "Community",
                "University"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of TRF contributions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TRFListResponse"
                }
              }
            }
          }
        },
        "x-openai-isConsequential": false,
        "x-ms-require-user-confirmation": false
      }
    },
...
  },
  "components": {
    "schemas": {
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "description": "Total matching records"
          },
          "limit": {
            "type": "integer",
            "description": "Records per page limit"
          },
          "offset": {
            "type": "integer",
            "description": "Record offset index"
          },
          "hasMore": {
            "type": "boolean",
            "description": "True if more records exist"
          }
        },
        "required": [
          "total",
          "limit",
          "offset",
          "hasMore"
        ]
      },
...
      "RotaryOpportunityClubItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "22000"
          },
          "name": {
            "type": "string",
            "example": "Rotary Club of Madurai West"
          },
          "district": {
            "type": "string",
            "example": "3000"
          },
          "zone": {
            "type": "string",
            "example": "Zone 5"
          },
          "members": {
            "type": "integer",
            "example": 45
          },
          "status": {
            "type": "string",
            "example": "Active"
          },
          "totalRotaractSponsored": {
            "type": "integer",
            "example": 0
          },
          "totalInteractSponsored": {
            "type": "integer",
            "example": 1
          }
        }
      },
      "RotaryOpportunityResponse": {
        "type": "object",
        "properties": {
          "opportunityType": {
            "type": "string",
            "example": "no_rotaract"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RotaryOpportunityClubItem"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMeta"
          },
          "total": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "offset": {
            "type": "integer"
          }
        }
      },
      "TRFContributionItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "91234"
          },
          "name": {
            "type": "string",
            "example": "Rotaract Club of Coimbatore Central"
          },
          "district": {
            "type": "string",
            "example": "3201"
          },
          "zone": {
            "type": "string",
            "example": "Zone 5"
          },
          "sponsorClubs": {
            "type": "string",
            "example": "Rotary Club of Coimbatore"
          },
          "annualFundUSD": {
            "type": "number",
            "example": 100
          },
          "polioPlusUSD": {
            "type": "number",
            "example": 50
          },
          "otherFundsUSD": {
            "type": "number",
            "example": 0
          },
          "totalContributionsUSD": {
            "type": "number",
            "example": 150
          }
        }
      },
      "TRFListResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TRFContributionItem"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMeta"
          },
          "total": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "offset": {
            "type": "integer"
          }
        }
      },
...
      "LeaderboardsResponse": {
        "type": "object",
        "properties": {
          "category": {
            "type": "string",
            "example": "largest_clubs"
          },
          "count": {
            "type": "integer",
            "example": 10
          },
          "data": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "largestClubs": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "topTRFClubs": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "highestArrearsClubs": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "topDistrictsByClubs": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "topDistrictsByMembers": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "newClubs": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
...
    }
  }
}